requirements.txt
# FastAPI & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
bcrypt==4.1.1
cryptography==41.0.7
passlib

# Proxmox API
requests==2.31.0
proxmoxer==1.3.1

# Ansible
ansible==2.10.7
ansible-runner==2.3.6

# Email
python-multipart==0.0.6
aiosmtplib==3.0.1

# Utilities
pydantic-extra-types==2.4.1
python-dateutil==2.8.2
pytz==2023.3

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
